requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
pandas
experta
groq
python-dotenv
